--- src/main/java/com/travelock/server/dto/BigBlockRequestDTO.java
@@ -1,15 +1,0 @@
-package com.travelock.server.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class BigBlockRequestDTO {
-    private String cityCode;
-    private String cityName;
-    private Long stateId;
-}
