--- src/main/java/com/travelock/server/dto/StateRequestDTO.java
@@ -1,13 +1,0 @@
-package com.travelock.server.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class StateRequestDTO {
-    private String stateCode;
-    private String stateName;
-}
