+++ src/main/java/com/travelock/server/repository/BigBlockRepository.java
@@ -7,11 +7,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface BigBlockRepository extends JpaRepository<BigBlock, Long> {
-
-//    // 빅블럭은 단일ID
-//    BigBlock findByStateAndCityCode(State state, String cityCode);
-//
-//
-//    // 특정 state 내에 속한 모든 목록 조회
-//   List<BigBlock> findByState(State state);
 }
